// src/models/releaseList.ts
export type ReleaseListInfo = {
  name: string;
  version: string;
  createName: number;
  createUser: string;
  publishTime: number;
  status: number;
  id: number;
  source: number;
  publishUser: string;
  appId: number;
  screenId: number;
  flowStatus: string;
  flowUrl: string;
  deleted: number;
  indexId?:number;
  experimentId?:number;
};